requirements.txt
asgiref==3.8.1
blinker==1.9.0
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==3.4.7
Deprecated==1.2.15
Django==4.2
dnspython==2.7.0
et_xmlfile==2.0.0
firebase-admin==6.6.0
Flask==3.1.0
Flask-Cors==5.0.0
gcloud==0.18.3
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
lxml==5.3.0
MarkupSafe==3.0.2
msgpack==1.1.0
numpy==1.23.5
oauth2client==4.1.3
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-semantic-conventions==0.50b0
packaging==24.2
Pillow==10.0.0
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.10.1
pymongo==4.10.1
<<<<<<< HEAD
pyparsing==3.2.0
=======
pyparsing==3.2.0    
>>>>>>> b2341e7889e447dffaa274c107459a9a67c2cca6
Pyrebase4==4.8.0
python-docx==1.1.2
python-dotenv==1.0.1
python-jwt==4.1.0
reportlab==4.2.5
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.20
Werkzeug==3.1.3
wrapt==1.17.0
zipp==3.21.0
